% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicrange_helpers.R
\name{bedpe2grl}
\alias{bedpe2grl}
\title{Convert Bedpe Table to GRangesList}
\usage{
bedpe2grl(bedpe, flip = FALSE, trim = TRUE, genome = NULL, sort = TRUE)
}
\description{
convert a bedpe in data frame format to a GRangesList
}

