% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_r_utilities.R
\name{colnames2}
\alias{colnames2}
\title{robust colnames

gives back character vector same number of columns of input regardless whether named or not}
\usage{
colnames2(x)
}
\arguments{
\item{str}{a path string}
}
\value{
a string with multiple parentheses replaced with a single parenthesis
}

