% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_r_utilities.R
\name{dodo.call2}
\alias{dodo.call2}
\title{dodo.call+

FUN can be an anonymous function call
dodo.call2({function(...) paste(..., collapse = " ")}, list(bstring1, bstring2))
can also omit the brackets}
\usage{
dodo.call2(FUN, args, use.names = T)
}
\arguments{
\item{FUN}{function}
}
\author{
Marcin Imielinski
}

