% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_r_utilities.R
\name{et}
\alias{et}
\title{shortcut for eval(parse(text = <string>))}
\usage{
et(txt, eval = TRUE, envir = parent.frame(), enclos = parent.frame(2))
}
\arguments{
\item{txt}{string to evaluate}
}
\author{
Kevin Hadi
}

