% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_r_utilities.R
\name{names2<-}
\alias{names2<-}
\title{robust name() assignment

similar to rlang::names2}
\usage{
names2(x, useempty = FALSE) <- value
}
\arguments{
\item{x}{vector}
}
\value{
x a vector with all names
}

