% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_r_utilities.R
\name{rn2col}
\alias{rn2col}
\title{alias for rownames_to_column

internal version that doesn't require library(tibble)}
\usage{
rn2col(.data, var = "rowname", keep.rownames = FALSE,
  asdf = as.data.frame, as.data.frame = FALSE)
}
\arguments{
\item{.data}{a data frame/table}
}
\value{
a data frame/table with rownames from a column
}

