% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_r_utilities.R
\name{setAllNames}
\alias{setAllNames}
\title{setAllNames

convenience function to set all names of a vector}
\usage{
setAllNames(vec, nm)
}
\arguments{
\item{vec}{vector}

\item{nm}{character}
}
\value{
named vector
}

